+++ crates/common/src/config/smtp/queue.rs
@@ -84,7 +84,8 @@ pub struct QueueOutboundTimeout {
 
 #[derive(Debug, Clone)]
 pub struct QueueThrottle {
-    pub outbound_concurrency: u64,
+    pub outbound_concurrency: usize,
+    pub local_concurrency: usize,
     pub sender: Vec<Throttle>,
     pub rcpt: Vec<Throttle>,
     pub host: Vec<Throttle>,
@@ -203,6 +204,7 @@ impl Default for QueueConfig {
             },
             throttle: QueueThrottle {
                 outbound_concurrency: 25,
+                local_concurrency: 10,
                 sender: Default::default(),
                 rcpt: Default::default(),
                 host: Default::default(),
@@ -387,8 +389,11 @@ fn parse_queue_throttle(config: &mut Config) -> QueueThrottle {
         rcpt: Vec::new(),
         host: Vec::new(),
         outbound_concurrency: config
-            .property_or_default::<u64>("queue.outbound.concurrency", "25")
+            .property_or_default::<usize>("queue.threads.remote", "25")
             .unwrap_or(25),
+        local_concurrency: config
+            .property_or_default::<usize>("queue.threads.local", "10")
+            .unwrap_or(10),
     };
 
     let all_throttles = parse_throttle(